Add files via upload
The current system is largely based on Python and requires an algorithm that  calculates the total cost of a customer's shopping cart, applies discounts if a valid  discount code is provided, and displays the final order total to the customer. However,  the company wants to keep track of all orders placed and gain insights into sales data  for future analysis.